move otlp_transformer changes over
